apps/api/app/core/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
redis[hiredis]==5.2.1
pydantic[email]==2.10.3
pydantic-settings==2.7.0
email-validator==2.2.0
python-jose[cryptography]==3.3.0
# passlib reemplazado por bcrypt directo en security.py
openai>=1.59.2
httpx==0.28.1
cryptography==44.0.0
python-multipart==0.0.20
qdrant-client==1.12.1
bcrypt==4.2.1
